file: seekjoiner cleanup
