+++ cmd/bee-file/main.go
@@ -17,7 +17,7 @@ import (
 	cmdfile "github.com/ethersphere/bee/cmd/internal/file"
 	"github.com/ethersphere/bee/pkg/collection/entry"
 	"github.com/ethersphere/bee/pkg/file"
-	"github.com/ethersphere/bee/pkg/file/seekjoiner"
+	"github.com/ethersphere/bee/pkg/file/joiner"
 	"github.com/ethersphere/bee/pkg/file/splitter"
 	"github.com/ethersphere/bee/pkg/logging"
 	"github.com/ethersphere/bee/pkg/storage"
@@ -59,8 +59,12 @@ func getEntry(cmd *cobra.Command, args []string) (err error) {
 	buf := bytes.NewBuffer(nil)
 	writeCloser := cmdfile.NopWriteCloser(buf)
 	limitBuf := cmdfile.NewLimitWriteCloser(writeCloser, limitMetadataLength)
-	j := seekjoiner.NewSimpleJoiner(store)
-	_, err = file.JoinReadAll(cmd.Context(), j, addr, limitBuf)
+	j, _, err := joiner.New(cmd.Context(), store, addr)
+	if err != nil {
+		return err
+	}
+
+	_, err = file.JoinReadAll(cmd.Context(), j, limitBuf)
 	if err != nil {
 		return err
 	}
@@ -70,8 +74,14 @@ func getEntry(cmd *cobra.Command, args []string) (err error) {
 		return err
 	}
 
+	j, _, err = joiner.New(cmd.Context(), store, e.Metadata())
+	if err != nil {
+		return err
+	}
+
 	buf = bytes.NewBuffer(nil)
-	_, err = file.JoinReadAll(cmd.Context(), j, e.Metadata(), buf)
+
+	_, err = file.JoinReadAll(cmd.Context(), j, buf)
 	if err != nil {
 		return err
 	}
@@ -117,7 +127,13 @@ func getEntry(cmd *cobra.Command, args []string) (err error) {
 		return err
 	}
 	defer outFile.Close()
-	_, err = file.JoinReadAll(cmd.Context(), j, e.Reference(), outFile)
+
+	j, _, err = joiner.New(cmd.Context(), store, e.Reference())
+	if err != nil {
+		return err
+	}
+
+	_, err = file.JoinReadAll(cmd.Context(), j, outFile)
 	return err
 }
