+++ pkg/api/bzz.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/ethersphere/bee/pkg/collection/entry"
 	"github.com/ethersphere/bee/pkg/file"
-	"github.com/ethersphere/bee/pkg/file/seekjoiner"
+	"github.com/ethersphere/bee/pkg/file/joiner"
 	"github.com/ethersphere/bee/pkg/jsonhttp"
 	"github.com/ethersphere/bee/pkg/manifest"
 	"github.com/ethersphere/bee/pkg/sctx"
@@ -53,9 +53,16 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 	toDecrypt := len(address.Bytes()) == 64
 
 	// read manifest entry
-	j := seekjoiner.NewSimpleJoiner(s.Storer)
+	j, _, err := joiner.New(ctx, s.Storer, address)
+	if err != nil {
+		logger.Debugf("bzz download: joiner manifest entry %s: %v", address, err)
+		logger.Errorf("bzz download: joiner %s", address)
+		jsonhttp.NotFound(w, nil)
+		return
+	}
+
 	buf := bytes.NewBuffer(nil)
-	_, err = file.JoinReadAll(ctx, j, address, buf)
+	_, err = file.JoinReadAll(ctx, j, buf)
 	if err != nil {
 		logger.Debugf("bzz download: read entry %s: %v", address, err)
 		logger.Errorf("bzz download: read entry %s", address)
@@ -72,8 +79,17 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// read metadata
+	j, _, err = joiner.New(ctx, s.Storer, e.Metadata())
+	if err != nil {
+		logger.Debugf("bzz download: joiner metadata %s: %v", address, err)
+		logger.Errorf("bzz download: joiner %s", address)
+		jsonhttp.NotFound(w, nil)
+		return
+	}
+
+	// read metadata
 	buf = bytes.NewBuffer(nil)
-	_, err = file.JoinReadAll(ctx, j, e.Metadata(), buf)
+	_, err = file.JoinReadAll(ctx, j, buf)
 	if err != nil {
 		logger.Debugf("bzz download: read metadata %s: %v", address, err)
 		logger.Errorf("bzz download: read metadata %s", address)
@@ -114,7 +130,7 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 				// index document exists
 				logger.Debugf("bzz download: serving path: %s", pathWithIndex)
 
-				s.serveManifestEntry(w, r, j, address, indexDocumentManifestEntry.Reference())
+				s.serveManifestEntry(w, r, address, indexDocumentManifestEntry.Reference())
 				return
 			}
 		}
@@ -154,7 +170,7 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 						// index document exists
 						logger.Debugf("bzz download: serving path: %s", pathWithIndex)
 
-						s.serveManifestEntry(w, r, j, address, indexDocumentManifestEntry.Reference())
+						s.serveManifestEntry(w, r, address, indexDocumentManifestEntry.Reference())
 						return
 					}
 				}
@@ -168,7 +184,7 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 						// error document exists
 						logger.Debugf("bzz download: serving path: %s", errorDocumentPath)
 
-						s.serveManifestEntry(w, r, j, address, errorDocumentManifestEntry.Reference())
+						s.serveManifestEntry(w, r, address, errorDocumentManifestEntry.Reference())
 						return
 					}
 				}
@@ -182,16 +198,26 @@ func (s *server) bzzDownloadHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// serve requested path
-	s.serveManifestEntry(w, r, j, address, me.Reference())
+	s.serveManifestEntry(w, r, address, me.Reference())
 }
 
-func (s *server) serveManifestEntry(w http.ResponseWriter, r *http.Request, j file.JoinSeeker, address, manifestEntryAddress swarm.Address) {
-	logger := tracing.NewLoggerWithTraceID(r.Context(), s.Logger)
-	ctx := r.Context()
+func (s *server) serveManifestEntry(w http.ResponseWriter, r *http.Request, address, manifestEntryAddress swarm.Address) {
+	var (
+		logger = tracing.NewLoggerWithTraceID(r.Context(), s.Logger)
+		ctx    = r.Context()
+		buf    = bytes.NewBuffer(nil)
+	)
 
 	// read file entry
-	buf := bytes.NewBuffer(nil)
-	_, err := file.JoinReadAll(ctx, j, manifestEntryAddress, buf)
+	j, _, err := joiner.New(ctx, s.Storer, manifestEntryAddress)
+	if err != nil {
+		logger.Debugf("bzz download: joiner read file entry %s: %v", address, err)
+		logger.Errorf("bzz download: joiner read file entry %s", address)
+		jsonhttp.NotFound(w, nil)
+		return
+	}
+
+	_, err = file.JoinReadAll(ctx, j, buf)
 	if err != nil {
 		logger.Debugf("bzz download: read file entry %s: %v", address, err)
 		logger.Errorf("bzz download: read file entry %s", address)
@@ -208,8 +234,16 @@ func (s *server) serveManifestEntry(w http.ResponseWriter, r *http.Request, j fi
 	}
 
 	// read file metadata
+	j, _, err = joiner.New(ctx, s.Storer, fe.Metadata())
+	if err != nil {
+		logger.Debugf("bzz download: joiner read file entry %s: %v", address, err)
+		logger.Errorf("bzz download: joiner read file entry %s", address)
+		jsonhttp.NotFound(w, nil)
+		return
+	}
+
 	buf = bytes.NewBuffer(nil)
-	_, err = file.JoinReadAll(ctx, j, fe.Metadata(), buf)
+	_, err = file.JoinReadAll(ctx, j, buf)
 	if err != nil {
 		logger.Debugf("bzz download: read file metadata %s: %v", address, err)
 		logger.Errorf("bzz download: read file metadata %s", address)
