+++ pkg/file/file.go
@@ -7,8 +7,6 @@ package file
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"io"
 
 	"github.com/ethersphere/bee/pkg/swarm"
@@ -19,10 +17,11 @@ type Reader interface {
 	io.ReaderAt
 }
 
-// JoinSeeker provides a Joiner that can seek.
-type JoinSeeker interface {
-	Join(ctx context.Context, address swarm.Address) (dataOut Reader, dataLength int64, err error)
-	Size(ctx context.Context, address swarm.Address) (dataLength int64, err error)
+// Joiner provides the inverse functionality of the Splitter.
+type Joiner interface {
+	Reader
+	// Size returns the span of the hash trie represented by the joiner's root hash.
+	Size() int64
 }
 
 // Splitter starts a new file splitting job.
@@ -32,69 +31,4 @@ type JoinSeeker interface {
 // When EOF is received and splitting is done, the resulting Swarm Address is returned.
 type Splitter interface {
 	Split(ctx context.Context, dataIn io.ReadCloser, dataLength int64, toEncrypt bool) (addr swarm.Address, err error)
-}
-
-// JoinReadAll reads all output from the provided SeekJoiner.
-func JoinReadAll(ctx context.Context, j JoinSeeker, addr swarm.Address, outFile io.Writer) (int64, error) {
-	r, l, err := j.Join(ctx, addr)
-	if err != nil {
-		return 0, err
-	}
-	// join, rinse, repeat until done
-	data := make([]byte, swarm.ChunkSize)
-	var total int64
-	for i := int64(0); i < l; i += swarm.ChunkSize {
-		cr, err := r.Read(data)
-		if err != nil {
-			return total, err
-		}
-		total += int64(cr)
-		cw, err := outFile.Write(data[:cr])
-		if err != nil {
-			return total, err
-		}
-		if cw != cr {
-			return total, fmt.Errorf("short wrote %d of %d for chunk %d", cw, cr, i)
-		}
-	}
-	if total != l {
-		return total, fmt.Errorf("received only %d of %d total bytes", total, l)
-	}
-	return total, nil
-}
-
-// SplitWriteAll writes all input from provided reader to the provided splitter
-func SplitWriteAll(ctx context.Context, s Splitter, r io.Reader, l int64, toEncrypt bool) (swarm.Address, error) {
-	chunkPipe := NewChunkPipe()
-	errC := make(chan error)
-	go func() {
-		buf := make([]byte, swarm.ChunkSize)
-		c, err := io.CopyBuffer(chunkPipe, r, buf)
-		if err != nil {
-			errC <- err
-		}
-		if c != l {
-			errC <- errors.New("read count mismatch")
-		}
-		err = chunkPipe.Close()
-		if err != nil {
-			errC <- err
-		}
-		close(errC)
-	}()
-
-	addr, err := s.Split(ctx, chunkPipe, l, toEncrypt)
-	if err != nil {
-		return swarm.ZeroAddress, err
-	}
-
-	select {
-	case err := <-errC:
-		if err != nil {
-			return swarm.ZeroAddress, err
-		}
-	case <-ctx.Done():
-		return swarm.ZeroAddress, ctx.Err()
-	}
-	return addr, nil
 }
