+++ pkg/file/file_test.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 
 	"github.com/ethersphere/bee/pkg/file"
+	"github.com/ethersphere/bee/pkg/file/joiner"
 	"github.com/ethersphere/bee/pkg/file/pipeline/builder"
-	"github.com/ethersphere/bee/pkg/file/seekjoiner"
 	test "github.com/ethersphere/bee/pkg/file/testing"
 	"github.com/ethersphere/bee/pkg/storage"
 	"github.com/ethersphere/bee/pkg/storage/mock"
@@ -45,7 +45,6 @@ func testSplitThenJoin(t *testing.T) {
 		dataIdx, _  = strconv.ParseInt(paramstring[1], 10, 0)
 		store       = mock.NewStorer()
 		p           = builder.NewPipelineBuilder(context.Background(), store, storage.ModePutUpload, false)
-		j           = seekjoiner.NewSimpleJoiner(store)
 		data, _     = test.GetVector(t, int(dataIdx))
 	)
 
@@ -59,7 +58,7 @@ func testSplitThenJoin(t *testing.T) {
 	}
 
 	// then join
-	r, l, err := j.Join(ctx, resultAddress)
+	r, l, err := joiner.New(ctx, store, resultAddress)
 	if err != nil {
 		t.Fatal(err)
 	}
