+++ pkg/file/io.go
@@ -6,9 +6,13 @@ package file
 
 import (
 	"bytes"
+	"context"
 	"errors"
+	"fmt"
 
 	"io"
+
+	"github.com/ethersphere/bee/pkg/swarm"
 )
 
 // simpleReadCloser wraps a byte slice in a io.ReadCloser implementation.
@@ -39,4 +43,67 @@ func (s *simpleReadCloser) Close() error {
 	}
 	s.closed = true
 	return nil
+}
+
+// JoinReadAll reads all output from the provided Joiner.
+func JoinReadAll(ctx context.Context, j Joiner, outFile io.Writer) (int64, error) {
+	l := j.Size()
+
+	// join, rinse, repeat until done
+	data := make([]byte, swarm.ChunkSize)
+	var total int64
+	for i := int64(0); i < l; i += swarm.ChunkSize {
+		cr, err := j.Read(data)
+		if err != nil {
+			return total, err
+		}
+		total += int64(cr)
+		cw, err := outFile.Write(data[:cr])
+		if err != nil {
+			return total, err
+		}
+		if cw != cr {
+			return total, fmt.Errorf("short wrote %d of %d for chunk %d", cw, cr, i)
+		}
+	}
+	if total != l {
+		return total, fmt.Errorf("received only %d of %d total bytes", total, l)
+	}
+	return total, nil
+}
+
+// SplitWriteAll writes all input from provided reader to the provided splitter
+func SplitWriteAll(ctx context.Context, s Splitter, r io.Reader, l int64, toEncrypt bool) (swarm.Address, error) {
+	chunkPipe := NewChunkPipe()
+	errC := make(chan error)
+	go func() {
+		buf := make([]byte, swarm.ChunkSize)
+		c, err := io.CopyBuffer(chunkPipe, r, buf)
+		if err != nil {
+			errC <- err
+		}
+		if c != l {
+			errC <- errors.New("read count mismatch")
+		}
+		err = chunkPipe.Close()
+		if err != nil {
+			errC <- err
+		}
+		close(errC)
+	}()
+
+	addr, err := s.Split(ctx, chunkPipe, l, toEncrypt)
+	if err != nil {
+		return swarm.ZeroAddress, err
+	}
+
+	select {
+	case err := <-errC:
+		if err != nil {
+			return swarm.ZeroAddress, err
+		}
+	case <-ctx.Done():
+		return swarm.ZeroAddress, ctx.Err()
+	}
+	return addr, nil
 }
