+++ pkg/file/joiner/joiner_test.go
@@ -2,24 +2,222 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package internal_test
+package joiner_test
 
 import (
 	"bytes"
 	"context"
+	"errors"
+	"fmt"
 	"io"
 	"io/ioutil"
 	mrand "math/rand"
 	"testing"
 	"time"
 
-	"github.com/ethersphere/bee/pkg/file/seekjoiner/internal"
+	"github.com/ethersphere/bee/pkg/encryption/store"
+	"github.com/ethersphere/bee/pkg/file/joiner"
+	"github.com/ethersphere/bee/pkg/file/pipeline/builder"
 	"github.com/ethersphere/bee/pkg/file/splitter"
 	filetest "github.com/ethersphere/bee/pkg/file/testing"
 	"github.com/ethersphere/bee/pkg/storage"
 	"github.com/ethersphere/bee/pkg/storage/mock"
 	"github.com/ethersphere/bee/pkg/swarm"
+	"gitlab.com/nolash/go-mockbytes"
 )
+
+func TestJoiner_ErrReferenceLength(t *testing.T) {
+	store := mock.NewStorer()
+	_, _, err := joiner.New(context.Background(), store, swarm.ZeroAddress)
+
+	if !errors.Is(err, storage.ErrReferenceLength) {
+		t.Fatalf("expected ErrReferenceLength %x but got %v", swarm.ZeroAddress, err)
+	}
+}
+
+// TestJoinerSingleChunk verifies that a newly created joiner returns the data stored
+// in the store when the reference is one single chunk.
+func TestJoinerSingleChunk(t *testing.T) {
+	store := mock.NewStorer()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// create the chunk to
+	mockAddrHex := fmt.Sprintf("%064s", "2a")
+	mockAddr := swarm.MustParseHexAddress(mockAddrHex)
+	mockData := []byte("foo")
+	mockDataLengthBytes := make([]byte, 8)
+	mockDataLengthBytes[0] = 0x03
+	mockChunk := swarm.NewChunk(mockAddr, append(mockDataLengthBytes, mockData...))
+	_, err := store.Put(ctx, storage.ModePutUpload, mockChunk)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// read back data and compare
+	joinReader, l, err := joiner.New(ctx, store, mockAddr)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if l != int64(len(mockData)) {
+		t.Fatalf("expected join data length %d, got %d", len(mockData), l)
+	}
+	joinData, err := ioutil.ReadAll(joinReader)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !bytes.Equal(joinData, mockData) {
+		t.Fatalf("retrieved data '%x' not like original data '%x'", joinData, mockData)
+	}
+}
+
+// TestJoinerDecryptingStore_NormalChunk verifies the the mock store that uses
+// the decrypting store manages to retrieve a normal chunk which is not encrypted
+func TestJoinerDecryptingStore_NormalChunk(t *testing.T) {
+	st := mock.NewStorer()
+	store := store.New(st)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// create the chunk to
+	mockAddrHex := fmt.Sprintf("%064s", "2a")
+	mockAddr := swarm.MustParseHexAddress(mockAddrHex)
+	mockData := []byte("foo")
+	mockDataLengthBytes := make([]byte, 8)
+	mockDataLengthBytes[0] = 0x03
+	mockChunk := swarm.NewChunk(mockAddr, append(mockDataLengthBytes, mockData...))
+	_, err := st.Put(ctx, storage.ModePutUpload, mockChunk)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// read back data and compare
+	joinReader, l, err := joiner.New(ctx, store, mockAddr)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if l != int64(len(mockData)) {
+		t.Fatalf("expected join data length %d, got %d", len(mockData), l)
+	}
+	joinData, err := ioutil.ReadAll(joinReader)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !bytes.Equal(joinData, mockData) {
+		t.Fatalf("retrieved data '%x' not like original data '%x'", joinData, mockData)
+	}
+}
+
+// TestJoinerWithReference verifies that a chunk reference is correctly resolved
+// and the underlying data is returned.
+func TestJoinerWithReference(t *testing.T) {
+	store := mock.NewStorer()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	// create root chunk and two data chunks referenced in the root chunk
+	rootChunk := filetest.GenerateTestRandomFileChunk(swarm.ZeroAddress, swarm.ChunkSize*2, swarm.SectionSize*2)
+	_, err := store.Put(ctx, storage.ModePutUpload, rootChunk)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	firstAddress := swarm.NewAddress(rootChunk.Data()[8 : swarm.SectionSize+8])
+	firstChunk := filetest.GenerateTestRandomFileChunk(firstAddress, swarm.ChunkSize, swarm.ChunkSize)
+	_, err = store.Put(ctx, storage.ModePutUpload, firstChunk)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	secondAddress := swarm.NewAddress(rootChunk.Data()[swarm.SectionSize+8:])
+	secondChunk := filetest.GenerateTestRandomFileChunk(secondAddress, swarm.ChunkSize, swarm.ChunkSize)
+	_, err = store.Put(ctx, storage.ModePutUpload, secondChunk)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// read back data and compare
+	joinReader, l, err := joiner.New(ctx, store, rootChunk.Address())
+	if err != nil {
+		t.Fatal(err)
+	}
+	if l != int64(swarm.ChunkSize*2) {
+		t.Fatalf("expected join data length %d, got %d", swarm.ChunkSize*2, l)
+	}
+
+	resultBuffer := make([]byte, swarm.ChunkSize)
+	n, err := joinReader.Read(resultBuffer)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if n != len(resultBuffer) {
+		t.Fatalf("expected read count %d, got %d", len(resultBuffer), n)
+	}
+	if !bytes.Equal(resultBuffer, firstChunk.Data()[8:]) {
+		t.Fatalf("expected resultbuffer %v, got %v", resultBuffer, firstChunk.Data()[:len(resultBuffer)])
+	}
+}
+
+func TestEncryptDecrypt(t *testing.T) {
+	var tests = []struct {
+		chunkLength int
+	}{
+		{10},
+		{100},
+		{1000},
+		{4095},
+		{4096},
+		{4097},
+		{1000000},
+	}
+
+	for _, tt := range tests {
+		t.Run(fmt.Sprintf("Encrypt %d bytes", tt.chunkLength), func(t *testing.T) {
+			store := mock.NewStorer()
+
+			g := mockbytes.New(0, mockbytes.MockTypeStandard).WithModulus(255)
+			testData, err := g.SequentialBytes(tt.chunkLength)
+			if err != nil {
+				t.Fatal(err)
+			}
+			ctx := context.Background()
+			pipe := builder.NewPipelineBuilder(ctx, store, storage.ModePutUpload, true)
+			testDataReader := bytes.NewReader(testData)
+			resultAddress, err := builder.FeedPipeline(ctx, pipe, testDataReader, int64(len(testData)))
+			if err != nil {
+				t.Fatal(err)
+			}
+			reader, l, err := joiner.New(context.Background(), store, resultAddress)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if l != int64(len(testData)) {
+				t.Fatalf("expected join data length %d, got %d", len(testData), l)
+			}
+
+			totalGot := make([]byte, tt.chunkLength)
+			index := 0
+			resultBuffer := make([]byte, swarm.ChunkSize)
+
+			for index < tt.chunkLength {
+				n, err := reader.Read(resultBuffer)
+				if err != nil && err != io.EOF {
+					t.Fatal(err)
+				}
+				copy(totalGot[index:], resultBuffer[:n])
+				index += n
+			}
+
+			if !bytes.Equal(testData, totalGot) {
+				t.Fatal("input data and output data does not match")
+			}
+		})
+	}
+}
 
 func TestSeek(t *testing.T) {
 	seed := time.Now().UnixNano()
@@ -80,7 +278,7 @@ func TestSeek(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			j, _, err := internal.NewSimpleJoiner(ctx, store, addr)
+			j, _, err := joiner.New(ctx, store, addr)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -358,7 +556,7 @@ func TestPrefetch(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			j, _, err := internal.NewSimpleJoiner(ctx, store, addr)
+			j, _, err := joiner.New(ctx, store, addr)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -378,8 +576,7 @@ func TestPrefetch(t *testing.T) {
 	}
 }
 
-// TestSimpleJoinerReadAt
-func TestSimpleJoinerReadAt(t *testing.T) {
+func TestJoinerReadAt(t *testing.T) {
 	store := mock.NewStorer()
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
@@ -406,7 +603,7 @@ func TestSimpleJoinerReadAt(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	j, _, err := internal.NewSimpleJoiner(ctx, store, rootChunk.Address())
+	j, _, err := joiner.New(ctx, store, rootChunk.Address())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -422,9 +619,9 @@ func TestSimpleJoinerReadAt(t *testing.T) {
 	}
 }
 
-// TestSimpleJoinerOneLevel tests the retrieval of two data chunks immediately
+// TestJoinerOneLevel tests the retrieval of two data chunks immediately
 // below the root chunk level.
-func TestSimpleJoinerOneLevel(t *testing.T) {
+func TestJoinerOneLevel(t *testing.T) {
 	store := mock.NewStorer()
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
@@ -451,7 +648,7 @@ func TestSimpleJoinerOneLevel(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	j, _, err := internal.NewSimpleJoiner(ctx, store, rootChunk.Address())
+	j, _, err := joiner.New(ctx, store, rootChunk.Address())
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -493,10 +690,10 @@ func TestSimpleJoinerOneLevel(t *testing.T) {
 	}
 }
 
-// TestSimpleJoinerTwoLevelsAcrossChunk tests the retrieval of data chunks below
+// TestJoinerTwoLevelsAcrossChunk tests the retrieval of data chunks below
 // first intermediate level across two intermediate chunks.
 // Last chunk has sub-chunk length.
-func TestSimpleJoinerTwoLevelsAcrossChunk(t *testing.T) {
+func TestJoinerTwoLevelsAcrossChunk(t *testing.T) {
 	store := mock.NewStorer()
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
@@ -543,7 +740,7 @@ func TestSimpleJoinerTwoLevelsAcrossChunk(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	j, _, err := internal.NewSimpleJoiner(ctx, store, rootChunk.Address())
+	j, _, err := joiner.New(ctx, store, rootChunk.Address())
 	if err != nil {
 		t.Fatal(err)
 	}
